BQ load updates
Fix issue with load bq tasks never being already complete. The complete_markers
implementation was missing, so I added it. We now query the markers before preparing
the task so we now if it was run before or not.

Update load bq so that chains can be excluded from loading. We found an issue with
"kroma" not having traces and so we need to exclude it from the 4337 dataset.