Add audit for trace counts
For each block num_traces should be >= num_transactions. We did not have this
audit at ingestion so sometimes we would pass through blocks that were missing
traces.

This issue was discovered while working on AA 4337 datasets. On kroma we stopped
getting traces on 2024-02-11.

We want to continue ingesting Kroma so we added a special cause in the new trace
count audit to let kroma always pass.