Improve mem use in blockbatch models by writing parquet from DuckDB
Avoiding the conversion to polars and going through the PartitionedWriteManager does not
allow streaming writes. If we write from duckdb we can let DuckDB decide if a streaming
write is possible. This improves memory use by 30% (9GB -> 6GB on an example batch for
the refined_traces model).

Removing an order by clause on an intermediate query further reduced memory use in the
same example from 6GB -> 3GB.

This change is important for backfills as we were frequently running OOM on some base
blockbatches from around 2024/10 where the trace parquet files are around ~180MB in size.