Support arbitrary output/input databases in blockbath load by date/chain
This will help with a better integration with transforms because we can write data into a transforms
database from blockbatch loading and then do more custom things in transforms.

Or, maybe more importantly, we can start something out first as a transforms and then seamlessly
migrate it over to blockbatch loading where it will have better data integrity and monitoring.